cleanup trace levels; fix typecasts, typos and formatting issues

+++ crypto/cmp/cmp_int.h
@@ -729,7 +729,7 @@ DECLARE_ASN1_FUNCTIONS(CMP_PROTECTEDPART)
 int CMP_ASN1_OCTET_STRING_set1(ASN1_OCTET_STRING **tgt,
                                const ASN1_OCTET_STRING *src);
 int CMP_ASN1_OCTET_STRING_set1_bytes(ASN1_OCTET_STRING **tgt,
-                                     const unsigned char *bytes, size_t len);
+                                     const unsigned char *bytes, int len);
 X509_EXTENSIONS *CMP_X509_EXTENSIONS_dup(const X509_EXTENSIONS *exts);
 /* from cmp_status.c */
 int CMP_ASN1_get_int(int func, const ASN1_INTEGER *a);
